[mypyc] Refactor: extract function IR from mypyc.ops
Work on mypyc/mypyc#716.
