+++ mypyc/emit.py
@@ -13,8 +13,9 @@ from mypyc.ops import (
     is_float_rprimitive, is_bool_rprimitive, is_int_rprimitive, is_short_int_rprimitive,
     is_list_rprimitive, is_dict_rprimitive, is_set_rprimitive, is_tuple_rprimitive,
     is_none_rprimitive, is_object_rprimitive, object_rprimitive, is_str_rprimitive,
-    FuncDecl, int_rprimitive, is_optional_type, optional_value_type
+    int_rprimitive, is_optional_type, optional_value_type
 )
+from mypyc.func_ir import FuncDecl
 from mypyc.class_ir import ClassIR, all_concrete_classes
 from mypyc.namegen import NameGenerator, exported_name
 from mypyc.sametype import is_same_type
