+++ mypyc/exceptions.py
@@ -12,9 +12,10 @@ only be placed at the end of a basic block.
 from typing import List, Optional
 
 from mypyc.ops import (
-    FuncIR, BasicBlock, LoadErrorValue, Return, Branch, RegisterOp,
-    ERR_NEVER, ERR_MAGIC, ERR_FALSE, NO_TRACEBACK_LINE_NO,
+    BasicBlock, LoadErrorValue, Return, Branch, RegisterOp, ERR_NEVER, ERR_MAGIC,
+    ERR_FALSE, NO_TRACEBACK_LINE_NO,
 )
+from mypyc.func_ir import FuncIR
 
 
 def insert_exception_handling(ir: FuncIR) -> None:
