+++ mypyc/irbuild/function.py
@@ -11,11 +11,14 @@ from mypy.nodes import (
 )
 from mypy.types import CallableType, get_proper_type
 from mypyc.ops import (
-    BasicBlock, FuncSignature, Value, FuncIR, RuntimeArg, object_rprimitive, FuncDecl,
+    BasicBlock, Value, object_rprimitive,
     Return, Call, SetAttr, LoadInt, Unreachable, RaiseStandardError,
     Environment, GetAttr, Branch, AssignmentTarget, TupleGet, Goto,
     int_rprimitive, RInstance, AssignmentTargetRegister, AssignmentTargetAttr, LoadStatic,
-    InitStatic, FUNC_CLASSMETHOD, FUNC_STATICMETHOD, FUNC_NORMAL
+    InitStatic
+)
+from mypyc.func_ir import (
+    FuncIR, FuncSignature, RuntimeArg, FuncDecl, FUNC_CLASSMETHOD, FUNC_STATICMETHOD, FUNC_NORMAL
 )
 from mypyc.class_ir import ClassIR, NonExtClassInfo
 from mypyc.primitives.misc_ops import (
