+++ mypyc/sametype.py
@@ -1,9 +1,9 @@
 """Same type check for RTypes."""
 
 from mypyc.ops import (
-    RType, RTypeVisitor, RInstance, RPrimitive, RTuple, RVoid,
-    FuncSignature, RUnion
+    RType, RTypeVisitor, RInstance, RPrimitive, RTuple, RVoid, RUnion
 )
+from mypyc.func_ir import FuncSignature
 
 
 def is_same_type(a: RType, b: RType) -> bool:
