+++ mypyc/test/test_analysis.py
@@ -9,7 +9,7 @@ from mypy.errors import CompileError
 from mypyc.common import TOP_LEVEL_NAME
 from mypyc import analysis
 from mypyc import exceptions
-from mypyc.ops import format_func
+from mypyc.func_ir import format_func
 from mypyc.test.testutil import (
     ICODE_GEN_BUILTINS, use_custom_builtins, MypycDataSuite, build_ir_for_single_file,
     assert_test_output,
